requirements.txt
allure-pytest==2.5.0
allure-python-commons==2.5.0
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.2.2
glob2==0.6
idna==2.7
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.2.0
parse==1.8.4
parse-type==0.4.2
pluggy==0.6.0
py==1.5.4
pycparser==2.18
PyMySQL==0.9.2
pytest==3.6.1
pytest-base-url==1.4.1
pytest-bdd==2.21.0
pytest-html==1.16.1
pytest-metadata==1.7.0
pytest-selenium==1.13.0
pytest-variables==1.7.1
requests==2.19.1
selenium==3.13.0
six==1.11.0
urllib3==1.23
